refactor: 유저예외 정렬, parse 예외 생성
