--- src/main/java/com/gamemarket/common/exception/utils/UtilsException.java
@@ -1,14 +1,0 @@
-package com.gamemarket.common.exception.utils;
-
-import lombok.Getter;
-
-@Getter
-public class UtilsException extends RuntimeException {
-
-    private final UtilsExceptionCode code;
-
-    public UtilsException(final UtilsExceptionCode code) {
-        this.code = code;
-    }
-
-}
