+++ src/main/java/com/gamemarket/common/utils/JsonUtils.java
@@ -2,8 +2,8 @@ package com.gamemarket.common.utils;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.gamemarket.common.exception.utils.UtilsException;
-import com.gamemarket.common.exception.utils.UtilsExceptionCode;
+import com.gamemarket.common.exception.parse.ParseException;
+import com.gamemarket.common.exception.parse.ParseExceptionCode;
 import lombok.experimental.UtilityClass;
 
 @UtilityClass
@@ -15,7 +15,7 @@ public class JsonUtils {
         try {
             return MAPPER.writeValueAsString(object);
         } catch (JsonProcessingException e) {
-            throw new UtilsException(UtilsExceptionCode.JSON_PARSER_EXCEPTION);
+            throw new ParseException(ParseExceptionCode.JSON_PARSER_EXCEPTION);
         }
     }
